+++ database/migrations/2016_08_08_095231_create_login_activities_table.php
@@ -2,6 +2,7 @@
 
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
+use Illuminate\Support\Facades\Schema;
 
 class CreateLoginActivitiesTable extends Migration
 {
@@ -12,14 +13,16 @@ class CreateLoginActivitiesTable extends Migration
      */
     public function up()
     {
-        Schema::create('login_activities', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('user_id')->unsigned()->index()->nullable();
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('set null');
-            $table->string('client')->nullable();
-            $table->string('ip_address', 64);
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('login_activities')) {
+            Schema::create('login_activities', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('user_id')->unsigned()->index()->nullable();
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('set null');
+                $table->string('client')->nullable();
+                $table->string('ip_address', 64);
+                $table->timestamps();
+            });
+        }
     }
 
     /**
