+++ database/migrations/2016_08_08_112912_create_transactions_table.php
@@ -2,6 +2,7 @@
 
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
+use Illuminate\Support\Facades\Schema;
 
 class CreateTransactionsTable extends Migration
 {
@@ -12,21 +13,23 @@ class CreateTransactionsTable extends Migration
      */
     public function up()
     {
-        Schema::create('transactions', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('sender_id')->unsigned()->index()->nullable();
-            $table->foreign('sender_id')->references('id')->on('users')->onDelete("set null");
-            $table->integer('receiver_id')->unsigned()->index()->nullable();
-            $table->foreign('receiver_id')->references('id')->on('users')->onDelete("set null");
-            $table->string('order_id', 80)->nullable();
-            $table->integer('type')->unsigned()->index()->nullable();
-            $table->foreign('type')->references('id')->on('transaction_types')->onDelete("set null");
-            $table->decimal('amount', 16, 2);
-            $table->integer('status')->unsigned()->index()->nullable();
-            $table->foreign('status')->references('id')->on('transaction_statuses')->onDelete("set null");
-            $table->string('otp', 12)->nullable();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('transactions')) {
+            Schema::create('transactions', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('sender_id')->unsigned()->index()->nullable();
+                $table->foreign('sender_id')->references('id')->on('users')->onDelete("set null");
+                $table->integer('receiver_id')->unsigned()->index()->nullable();
+                $table->foreign('receiver_id')->references('id')->on('users')->onDelete("set null");
+                $table->string('order_id', 80)->nullable();
+                $table->integer('type')->unsigned()->index()->nullable();
+                $table->foreign('type')->references('id')->on('transaction_types')->onDelete("set null");
+                $table->decimal('amount', 16, 2);
+                $table->integer('status')->unsigned()->index()->nullable();
+                $table->foreign('status')->references('id')->on('transaction_statuses')->onDelete("set null");
+                $table->string('otp', 12)->nullable();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
