+++ database/migrations/2016_08_08_115158_create_histories_table.php
@@ -2,6 +2,7 @@
 
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
+use Illuminate\Support\Facades\Schema;
 
 class CreateHistoriesTable extends Migration
 {
@@ -12,18 +13,20 @@ class CreateHistoriesTable extends Migration
      */
     public function up()
     {
-        Schema::create('histories', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('transaction_id')->unsigned()->index()->nullable();
-            $table->foreign('transaction_id')->references('id')->on('transactions')->onDelete('set null');
-            $table->integer('user_id')->unsigned()->index()->nullable();
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('set null');
-            $table->string('description');
-            $table->decimal('debit', 16, 2);
-            $table->decimal('credit', 16, 2);
-            $table->decimal('current_balance', 16, 2);
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('histories')) {
+            Schema::create('histories', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('transaction_id')->unsigned()->index()->nullable();
+                $table->foreign('transaction_id')->references('id')->on('transactions')->onDelete('set null');
+                $table->integer('user_id')->unsigned()->index()->nullable();
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('set null');
+                $table->string('description');
+                $table->decimal('debit', 16, 2);
+                $table->decimal('credit', 16, 2);
+                $table->decimal('current_balance', 16, 2);
+                $table->timestamps();
+            });
+        }
     }
 
     /**
