+++ database/migrations/2016_12_27_184835_create_recharge_plans_table.php
@@ -13,17 +13,19 @@ class CreateRechargePlansTable extends Migration
      */
     public function up()
     {
-        Schema::create('recharge_plans', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('operator_id')->unsigned()->index();
-            $table->foreign('operator_id')->references('id')->on('operators')->onDelete('cascade');
-            $table->string('logo', 120);
-            $table->decimal('amount', 10, 2);
-            $table->string('validity', 120);
-            $table->boolean('status')->default(true);
-            $table->json('eligibility')->nullable();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('recharge_plans')) {
+            Schema::create('recharge_plans', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('operator_id')->unsigned()->index();
+                $table->foreign('operator_id')->references('id')->on('operators')->onDelete('cascade');
+                $table->string('logo', 120);
+                $table->decimal('amount', 10, 2);
+                $table->string('validity', 120);
+                $table->boolean('status')->default(true);
+                $table->json('eligibility')->nullable();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
