+++ database/migrations/2016_12_28_155846_create_data_bundles_table.php
@@ -13,18 +13,20 @@ class CreateDataBundlesTable extends Migration
      */
     public function up()
     {
-        Schema::create('data_bundles', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('operator_id')->unsigned()->index()->nullable();
-            $table->foreign('operator_id')->references('id')->on('operators')->onDelete('set null');
-            $table->string('logo', 120);
-            $table->string('bundle_size', 32);
-            $table->decimal('price', 10, 2);
-            $table->string('top_up_profile', 64);
-            $table->boolean('status')->default(true);
-            $table->json('eligibility')->nullable();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('bundles')) {
+            Schema::create('bundles', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('operator_id')->unsigned()->index()->nullable();
+                $table->foreign('operator_id')->references('id')->on('operators')->onDelete('set null');
+                $table->string('logo', 120);
+                $table->string('bundle_size', 32);
+                $table->decimal('price', 10, 2);
+                $table->string('top_up_profile', 64);
+                $table->boolean('status')->default(true);
+                $table->json('eligibility')->nullable();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -34,6 +36,6 @@ class CreateDataBundlesTable extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('data_bundles');
+        Schema::dropIfExists('bundles');
     }
 }
