+++ database/migrations/2017_01_15_121743_create_activity_logs_table.php
@@ -13,16 +13,18 @@ class CreateActivityLogsTable extends Migration
      */
     public function up()
     {
-        Schema::create('activity_logs', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('user_id')->unsigned()->index()->nullable();
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('set null');
-            $table->enum('account_level', [1, 2, 3, 4, 5])->comment('1 = Admin, 2 = Merchant, 3 = Agent, 4 = Personal, 5 = Others');
-            $table->string('activity', 120);
-            $table->string('client', 160)->nullable();
-            $table->string('ip_address', 64);
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('activity_logs')) {
+            Schema::create('activity_logs', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('user_id')->unsigned()->index()->nullable();
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('set null');
+                $table->enum('account_level', [1, 2, 3, 4, 5])->comment('1 = Admin, 2 = Merchant, 3 = Agent, 4 = Personal, 5 = Others');
+                $table->string('activity', 120);
+                $table->string('client', 160)->nullable();
+                $table->string('ip_address', 64);
+                $table->timestamps();
+            });
+        }
     }
 
     /**
