+++ database/migrations/2017_05_15_131937_create_otps_table.php
@@ -13,13 +13,15 @@ class CreateOtpsTable extends Migration
      */
     public function up()
     {
-        Schema::create('otps', function (Blueprint $table) {
-            $table->integer('user_id')->unsigned()->index();
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-            $table->string('otp', 12)->index();
-            $table->string('remarks', 32)->index()->nullable();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('otps')) {
+            Schema::create('otps', function (Blueprint $table) {
+                $table->integer('user_id')->unsigned()->index();
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+                $table->string('otp', 12)->index();
+                $table->string('remarks', 32)->index()->nullable();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
