+++ database/migrations/2017_05_18_112426_create_account_migrations_table.php
@@ -13,15 +13,17 @@ class CreateAccountMigrationsTable extends Migration
      */
     public function up()
     {
-        Schema::create('account_migrations', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('user_id')->unsigned()->index();
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-            $table->string('old_mobile_no', 32);
-            $table->string('new_mobile_no', 32);
-            $table->boolean('status')->default(false);
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('account_migrations')) {
+            Schema::create('account_migrations', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('user_id')->unsigned()->index();
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+                $table->string('old_mobile_no', 32);
+                $table->string('new_mobile_no', 32);
+                $table->boolean('status')->default(false);
+                $table->timestamps();
+            });
+        }
     }
 
     /**
