+++ database/migrations/2018_03_15_180537_create_promotional_offers_table.php
@@ -13,22 +13,24 @@ class CreatePromotionalOffersTable extends Migration
      */
     public function up()
     {
-        Schema::create('promotional_offers', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('promotion_id')->unsigned()->index()->nullable();
-            $table->foreign('promotion_id')->references('id')->on('promotions')->onDelete('set null');
-            $table->integer('commission_id')->unsigned()->index()->nullable();
-            $table->foreign('commission_id')->references('id')->on('commissions')->onDelete('set null');
-            $table->string('title', 32);
-            $table->enum('disbursement_to', ['Sender', 'Receiver'])->default('Sender');
-            $table->decimal('amount', 8, 2)->default(0);
-            $table->enum('percentage_or_fixed', ['f', 'p'])->default('f');
-            $table->string('avail_limit', 3)->comment("1 for Once, x for x times, I for Infinite");
-            $table->integer('created_by')->unsigned()->index()->nullable();
-            $table->foreign('created_by')->references('id')->on('users')->onDelete('set null');
-            $table->boolean('status')->default(false);
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('promotional_offers')) {
+            Schema::create('promotional_offers', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('promotion_id')->unsigned()->index()->nullable();
+                $table->foreign('promotion_id')->references('id')->on('promotions')->onDelete('set null');
+                $table->integer('commission_id')->unsigned()->index()->nullable();
+                $table->foreign('commission_id')->references('id')->on('commissions')->onDelete('set null');
+                $table->string('title', 32);
+                $table->enum('disbursement_to', ['Sender', 'Receiver'])->default('Sender');
+                $table->decimal('amount', 8, 2)->default(0);
+                $table->enum('percentage_or_fixed', ['f', 'p'])->default('f');
+                $table->string('avail_limit', 3)->comment("1 for Once, x for x times, I for Infinite");
+                $table->integer('created_by')->unsigned()->index()->nullable();
+                $table->foreign('created_by')->references('id')->on('users')->onDelete('set null');
+                $table->boolean('status')->default(false);
+                $table->timestamps();
+            });
+        }
     }
 
     /**
